+++ src/common/token/claims_test.go
@@ -2,41 +2,9 @@ package token
 
 import (
 	"github.com/goharbor/harbor/src/common/rbac"
-	"github.com/goharbor/harbor/src/common/utils/test"
-	"github.com/goharbor/harbor/src/core/config"
 	"github.com/stretchr/testify/assert"
-	"os"
-	"path"
-	"runtime"
 	"testing"
 )
-
-func TestMain(m *testing.M) {
-	_, f, _, ok := runtime.Caller(0)
-	if !ok {
-		panic("Failed to get current directory")
-	}
-	keyPath := path.Join(path.Dir(f), "test/private_key.pem")
-	os.Setenv("TOKEN_PRIVATE_KEY_PATH", keyPath)
-
-	server, err := test.NewAdminserver(nil)
-	if err != nil {
-		panic(err)
-	}
-	defer server.Close()
-
-	if err := os.Setenv("ADMINSERVER_URL", server.URL); err != nil {
-		panic(err)
-	}
-	if err := config.Init(); err != nil {
-		panic(err)
-	}
-
-	result := m.Run()
-	if result != 0 {
-		os.Exit(result)
-	}
-}
 
 func TestValid(t *testing.T) {
